+++ src/debugger.cpp
@@ -607,7 +607,7 @@ void debugger_t::process_command(char *c)
 			if ((i & 0b111) == 0) {
 				terminal->printf("\n %02x ", i & 0xff);
 			}
-			terminal->bg_color = system->core->vdc->palette[i & 0xff];
+			terminal->bg_color = system->core->vdc->crt_palette[i & 0xff];
 			terminal->printf("  ");
 			terminal->bg_color = (PUNCH_BLUE & 0x00ffffff) | 0xe0000000;
 		}
@@ -619,7 +619,7 @@ void debugger_t::process_command(char *c)
 			terminal_graphics_colors[3]
 		);
 		for (int i=0; i<4; i++) {
-			terminal->bg_color = system->core->vdc->palette[terminal_graphics_colors[i]];
+			terminal->bg_color = system->core->vdc->crt_palette[terminal_graphics_colors[i]];
 			terminal->printf("    ");
 		}
 		terminal->bg_color = (PUNCH_BLUE & 0x00ffffff) | 0xe0000000;
@@ -790,7 +790,7 @@ void debugger_t::memory_binary_dump(uint32_t address)
 		terminal->printf("%c%c ", res[i] & 0b10 ? '1' : '.', res[i] & 0b1 ? '1' : '.');
 	}
 	for (int i=0; i<8; i++) {
-		terminal->bg_color = system->core->vdc->palette[terminal_graphics_colors[res[i]]];
+		terminal->bg_color = system->core->vdc->crt_palette[terminal_graphics_colors[res[i]]];
 		terminal->printf(" ");
 	}
 	terminal->bg_color = (PUNCH_BLUE & 0x00ffffff) | 0xe0000000;
@@ -959,7 +959,7 @@ void debugger_t::enter_dgc_line(char *buffer)
 		terminal->printf("\n    ");
 		//for (int i=0; i<16; i++) terminal->cursor_right();
 		for (int i=0; i<4; i++) {
-			terminal->bg_color = system->core->vdc->palette[terminal_graphics_colors[i]];
+			terminal->bg_color = system->core->vdc->crt_palette[terminal_graphics_colors[i]];
 			terminal->printf("    ");
 		}
 		terminal->bg_color = (PUNCH_BLUE & 0x00ffffff) | 0xe0000000;